+++ Assets/_Scripts/_GradientNoise/ValueGeneration/ValueGenerator.cs
@@ -1,55 +1,11 @@
 using System;
 using _Scripts._GradientNoise.OpenSimplex;
+using _Scripts.ScriptableObjects;
+using _Scripts.ScriptableObjects.ValueGenerationSettings;
 using UnityEngine;
 
 namespace _Scripts._GradientNoise.ValueGeneration
 {
-    /**
-     * The type of Noise which should be used.
-     */
-    public enum NoiseType
-    {
-        OpenSimplex,
-        Perlin,
-    }
-    
-    /**
-     * This class stores the parameters for each generation step.
-     */
-    [Serializable] // With this it can be shown in the Inspector
-    public class ValueGenerationSettings
-    {
-        // General
-        [Header("General")] public NoiseType noiseType = NoiseType.Perlin;
-        [Range(0, 100)] public int percentage = 45;
-    
-        // Seed
-        [Header("Seed")] public bool useRandomSeed = true;
-        private bool _seedLocked = false;
-        [SerializeField] private string seed = "Hello World!";
-        [Range(1000.0f, 1000000.0f)] public float seedScale = 100000.0f;
-    
-        // Gradient noise settings
-        [Header("Gradient Noise")] [Range(0.0f, 1.0f)]
-        public float noiseScale = 0.033f;
-
-        [Header("For Mountain Generation")] public int stonePercentage = 10;
-        
-        // Seed can only be changed if there is no seed
-        public void SetSeed(string inSeed)
-        {
-            if (_seedLocked) return;
-    
-            seed = inSeed;
-            _seedLocked = true;
-        }
-    
-        public string GetSeed()
-        {
-            return seed;
-        }
-    }
-
     /**
      * This class generates a value per cell (x-, y-coordinate pair)
      */